GeoAlert-Pipeline/kafka-consumer/requirements.txt
kafka-python==2.0.2
pandas==2.2.3
psycopg2-binary==2.9.9
